function openFormPopup
