+++ index.html
@@ -254,7 +254,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="./nouislider/nouislider.js"></script>
   <script src="./js/main.js" type="module"></script>
 </body>
