+++ js/form.js
@@ -1,22 +1,20 @@
+import './../nouislider/nouislider.js';
+import './filters.js';
+import './scale.js';
+import './slider.js';
+import './input-validate.js';
+
 import { isEscape } from './utils.js';
-import { showErrorMessage, showSuccessMessage } from './message.js';
 import { sendData } from './api.js';
-import { scaleControllSmallerButton, scaleControllBiggerButton, scaleControllValue, onScaleSmallerClick, onScaleBiggerClick, scaleValueHidden } from './scale.js';
 import { fileChooser } from './picture-preview.js';
-import { imgUploadPreview, sliderElementBlock, effectLevelValue } from './slider.js';
+import { showErrorMessage, showSuccessMessage } from './message.js';
+import { hashtagValidate, commentValidate, hashtags, commentField, imageLoad, onHashtagsTextInput, commentTextInput } from './input-validate.js';
+import { scaleControllSmallerButton, scaleControllBiggerButton, scaleControllValue, onScaleSmallerClick, onScaleBiggerClick, scaleValueHidden } from './scale.js';
+import { onEffectChange, imgUploadPreview, sliderElementBlock, effectLevelValue } from './slider.js';
 
 const formUploadImage = document.querySelector('.img-upload__form');
-const imageLoad = document.querySelector('.img-upload__overlay');
 const modalView = document.querySelector('body');
 const buttonModalClose = document.querySelector('.img-upload__cancel');
-
-const hashtags = imageLoad.querySelector('.text__hashtags');
-const regExp = /^#[A-Za-zА-Яа-яЁё0-9]{1,19}$/;
-const MAX_HASHTAG_QUANTITY = 5;
-
-const COMMENT_MAX_LENGTH = 140;
-const commentField = imageLoad.querySelector('.text__description');
-
 
 //слушатель изменения значения кнопкок scale
 const scaleChange = () => {
@@ -26,11 +24,18 @@ const scaleChange = () => {
 
 //слушатель изменения значения поля #upload-file)
 formUploadImage.addEventListener('change', () => {
+  openFormPopup();
+});
+
+//функция открытия модального окна
+function openFormPopup() {
   imageLoad.classList.remove('hidden');
   modalView.classList.add('modal-open');
   scaleChange();
-  formUploadImage.value = '';
-});
+  hashtagValidate();
+  commentValidate();
+  onEffectChange(); //не работает?/////////////////////////////////////////
+}
 
 //функция закрытия модального окна
 function closeFormPopup() {
@@ -39,15 +44,17 @@ function closeFormPopup() {
   //обработчики событий
   scaleControllSmallerButton.removeEventListener('click', onScaleSmallerClick);
   scaleControllBiggerButton.removeEventListener('click', onScaleBiggerClick);
+  hashtags.removeEventListener('input', onHashtagsTextInput);
+  commentField.addEventListener('input', commentTextInput);
   //Очистка полей
   fileChooser.value = '';
   scaleValueHidden.value = '100';
   scaleControllValue.value = '100%';
   imgUploadPreview.style.transform = 'scale(1)';
-  //фильтры
+  //фильтры (не работает)//////////////////////////////////////////////////
   sliderElementBlock.classList.add('hidden');
-  imgUploadPreview.style.filter = 'none';
-  effectLevelValue.value = 'none';
+  imgUploadPreview.style.filter = '';
+  effectLevelValue.value = '';
   // вводимый текст
   hashtags.value = '';
   commentField.value = '';
@@ -83,74 +90,4 @@ const setImgUploadFormSubmit = (onSuccess) => {
 
 setImgUploadFormSubmit(closeFormPopup);
 
-
-//валидация хештегов
-
-const onHashtagsTextInput = () => {
-  hashtags.value = hashtags.value.replaceAll('  ', ' ');
-
-  const hashtagsArr = hashtags.value.split(' ');
-  const invalidHashtagsArr = [];
-
-  if (hashtagsArr[0] === '') {
-    hashtagsArr.shift();
-  }
-  if (hashtagsArr[hashtagsArr.length - 1] === '') {
-    hashtagsArr.pop();
-  }
-  hashtagsArr.forEach((hashtag) => {
-    if (!hashtag.match(regExp)) {
-      invalidHashtagsArr.push(hashtag);
-    }
-  });
-
-  for (let i = 0; i < hashtagsArr.length; i++) {
-    hashtagsArr[i] = hashtagsArr[i].toLowerCase();
-  }
-
-  const duplicateHashtagsArr = hashtagsArr.filter((hashtag, index, arr) => arr.indexOf(hashtag) !== index);
-
-  if (duplicateHashtagsArr && duplicateHashtagsArr.length !== 0) {
-    hashtags.setCustomValidity(`Пожалуйста, удалите повторяющиеся хэш-теги: ${duplicateHashtagsArr.join(', ')}`);
-    hashtags.style.borderColor = '#FF5F49';
-  } else if (hashtagsArr.length > MAX_HASHTAG_QUANTITY) {
-    hashtags.setCustomValidity(`Нельзя указывать больше ${MAX_HASHTAG_QUANTITY} хэш-тегов. Просьба удалить лишние ${hashtagsArr.length - MAX_HASHTAG_QUANTITY}`);
-    hashtags.style.borderColor = '#FF5F49';
-  } else if (invalidHashtagsArr.length !== 0) {
-    hashtags.setCustomValidity(`Некорректно введен хэш-тег: ${invalidHashtagsArr.join(', ')}`);
-    hashtags.style.borderColor = '#FF5F49';
-  } else {
-    hashtags.setCustomValidity('');
-    hashtags.style.borderColor = '';
-  }
-  hashtags.reportValidity();
-};
-
-hashtags.addEventListener('input', onHashtagsTextInput);
-
-//отключает закрытие окна при фокусе в инпуте
-hashtags.addEventListener('keydown', (evt) => {
-  evt.stopPropagation();
-});
-
-//проверка поля ввода комментария
-const commentTextInput = () => {
-  const valueLength = commentField.value.length;
-  if (valueLength > COMMENT_MAX_LENGTH) {
-    commentField.setCustomValidity(`Максимальная длина комментария 140 символов. Удалите лишние ${valueLength - COMMENT_MAX_LENGTH} симв.`);
-    commentField.style.borderColor = '#FF5F49';
-  } else {
-    commentField.setCustomValidity('');
-    commentField.style.borderColor = '';
-  }
-  commentField.reportValidity();
-};
-
-commentField.addEventListener('input', commentTextInput);
-
-//отключает закрытие окна при фокусе в инпуте
-commentField.addEventListener('keydown', (evt) => {
-  evt.stopPropagation();
-});
-
-export { formUploadImage };
+export { formUploadImage, imageLoad };
