+++ js/main.js
@@ -3,9 +3,6 @@ import './create-big-pictures.js';
 import './show-comments.js';
 import './big-pictures-popup.js';
 import './form.js';
-import './scale.js';
-import './slider.js';
 import './api.js';
 import './picture-preview.js';
-import './filters.js';
 import './message.js';
